+++ Packages/com.unity.services.mediation/Editor/Settings/Adapters/UnityAdsSettings.cs
@@ -1,8 +1,6 @@
 using System;
 using UnityEditor;
 using UnityEngine;
-using System.Collections.Generic;
-using UnityEditor.UIElements;
 using UnityEngine.UIElements;
 
 namespace Unity.Services.Mediation.Settings.Editor