+++ Packages/com.unity.services.mediation/Editor/Settings/MediationAdapterSettings.cs
@@ -4,9 +4,7 @@ using System.Linq;
 using Unity.Services.Mediation.Adapters.Editor;
 using UnityEditor;
 using UnityEditor.Advertisements;
-using UnityEditor.SettingsManagement;
 using UnityEngine;
-using UnityEditor.UIElements;
 using UnityEngine.UIElements;
 
 
@@ -77,7 +75,6 @@ namespace Unity.Services.Mediation.Settings.Editor
         /// </summary>
         static void Refresh()
         {
-            Initialize();
             var installedAdapters = MediationSdkInfo.GetInstalledAdapters();
             var changed = false;
 
@@ -157,14 +154,6 @@ namespace Unity.Services.Mediation.Settings.Editor
         {
             VisualElement rootElement = new VisualElement();
 
-#if !ENABLE_EDITOR_GAME_SERVICES && UNITY_2020_1_OR_NEWER
-            if (string.IsNullOrEmpty(CloudProjectSettings.projectId))
-            {
-                SettingsService.OpenProjectSettings("Project/Services");
-                return rootElement;
-            }
-#endif
-
             VisualTreeAsset settingsTemplate = AssetDatabase.LoadAssetAtPath<VisualTreeAsset>(k_SettingsTemplate);
             VisualTreeAsset adapterTemplate  = AssetDatabase.LoadAssetAtPath<VisualTreeAsset>(k_AdapterTemplate);
 
@@ -226,8 +215,8 @@ namespace Unity.Services.Mediation.Settings.Editor
 #if !ENABLE_EDITOR_GAME_SERVICES
             MediationEditorService.RefreshGameId();
 #endif
-            rootElement.Q<TextElement>("android-game-id").text = AdvertisementSettings.GetGameId(RuntimePlatform.IPhonePlayer);
-            rootElement.Q<TextElement>("ios-game-id").text     = AdvertisementSettings.GetGameId(RuntimePlatform.Android);
+            rootElement.Q<TextElement>("android-game-id").text = AdvertisementSettings.GetGameId(RuntimePlatform.Android);
+            rootElement.Q<TextElement>("ios-game-id").text     = AdvertisementSettings.GetGameId(RuntimePlatform.IPhonePlayer);
 
             return rootElement;
         }
@@ -275,9 +264,7 @@ namespace Unity.Services.Mediation.Settings.Editor
         /// </summary>
         static void RefreshInstallButton(string adapterIdentifier)
         {
-            bool isInstalled = !string.IsNullOrEmpty(s_AdapterSettings.
-                FirstOrDefault(adapterSettings => adapterSettings.AdapterId == adapterIdentifier)?.
-                InstalledVersion.value);
+            bool isInstalled = IsAdapterInstalled(adapterIdentifier);
 
             s_AdapterInstallButton[adapterIdentifier].text = isInstalled ? k_Uninstall : k_Install;
         }
@@ -289,9 +276,7 @@ namespace Unity.Services.Mediation.Settings.Editor
         {
             if (adapterIdentifier != default)
             {
-                bool isInstalled = !string.IsNullOrEmpty(s_AdapterSettings.
-                    FirstOrDefault(adapterSettings => adapterSettings.AdapterId == adapterIdentifier)?.
-                    InstalledVersion.value);
+                bool isInstalled = IsAdapterInstalled(adapterIdentifier);
 
                 if (!isInstalled)
                 {
@@ -304,6 +289,13 @@ namespace Unity.Services.Mediation.Settings.Editor
 
                 RefreshInstallButton(adapterIdentifier);
             }
+        }
+
+        static bool IsAdapterInstalled(string adapterIdentifier)
+        {
+            return !string.IsNullOrEmpty(s_AdapterSettings.
+                FirstOrDefault(adapterSettings => adapterSettings.AdapterId == adapterIdentifier)?.
+                InstalledVersion.value);
         }
     }
 }